Update uffminimize.py
Added a check to expand the number of iterations that the compounds go through in the UFFOptimizeMolecule procedure if it failed to converge the first time around. I set the maxIters up from the default of 200 to 1000. This still runs reasonably quickly, and can help in the generation of improper compounds.